refactor: remove unused imports and streamline session management in database module
