+++ sdk/longlink/database/__init__.py
@@ -1,3 +1,2 @@
 from . import audit
 from .base import session
-from longlink.shared.models import Audit
