+++ sdk/longlink/database/base.py
@@ -86,37 +86,27 @@ def create_engine(env: Envs) -> AsyncEngine:
 @asynccontextmanager
 async def session() -> AsyncGenerator[AsyncSession, None]:
     """Yield an Application database session."""
-
-    # Open one session from the lazily initialized session factory.
-    session_maker = await get_session_maker()
-    async with session_maker() as session:
-        yield session
-
-
-async def get_session_maker() -> async_sessionmaker[AsyncSession]:
-    """Return a SQLModel async sessionmaker instance."""
     global Session
 
-    # Reuse the cached session factory once initialized.
-    if Session is not None:
-        return Session
-
     # Initialize the engine lazily when sessions are requested first.
-    engine = create_engine(Envs())
+    if Session is None:
+        engine = create_engine(Envs())
 
-    # Auto-create tables for SQLite only.
-    if str(engine.url).startswith("sqlite+"):
+        # Auto-create tables for SQLite only.
+        if str(engine.url).startswith("sqlite+"):
 
-        # Create tables through a transactional SQLite connection.
-        async with engine.begin() as conn:
-            await conn.run_sync(database_metadata.create_all)
-    else:
+            # Create tables through a transactional SQLite connection.
+            async with engine.begin() as conn:
+                await conn.run_sync(database_metadata.create_all)
+        else:
 
-        # Verify non-SQLite connections before exposing the session factory.
-        async with engine.connect():
-            pass
+            # Verify non-SQLite connections before exposing the session factory.
+            async with engine.connect():
+                pass
 
-    # Cache the session factory after the engine connection succeeds.
-    Session = async_sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
+        # Cache the session factory after the engine connection succeeds.
+        Session = async_sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
 
-    return Session
+    # Open one session from the lazily initialized session factory.
+    async with Session() as session:
+        yield session
