--- sdk/longlink/utils/__init__.py
@@ -1,1 +1,0 @@
-from longlink.utils.settings import Envs
